Change float to double and increase exponent for phe encoder

+++ src/executor/algorithm/model.cc
@@ -9,12 +9,12 @@ Model::Model() {
   testing_accuracy = 0.0;
 }
 
-Model::Model(std::vector<std::vector<float> > m_training_data,
-             std::vector<std::vector<float> > m_testing_data,
-             std::vector<float> m_training_labels,
-             std::vector<float> m_testing_labels,
-             float m_training_accuracy,
-             float m_testing_accuracy) {
+Model::Model(std::vector<std::vector<double> > m_training_data,
+             std::vector<std::vector<double> > m_testing_data,
+             std::vector<double> m_training_labels,
+             std::vector<double> m_testing_labels,
+             double m_training_accuracy,
+             double m_testing_accuracy) {
   training_data = m_training_data;
   training_labels = m_training_labels;
   training_accuracy = m_training_accuracy;
